Adding a delete CNUC feature

See merge request mike-ensor/abm-consumer-edge-ansible!17

+++ scripts/delete-cloud-gce-baseline.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-#TODO: run a script that unbundles ABM from GKE connect/kub
+echo "Looking for instances..."
 
 INSTANCES=()
 # By default, remove all GCE instances
@@ -14,13 +14,25 @@ if [[ ! -z "$1" ]]; then
     INSTANCES+=( "${instance_name}" )
 else
     # get list of all CNUCs in the project
-    INSTANCES=( $(gcloud compute instances list --format="value(name)") )
+    LABELS="labels.type=abm"
+    INSTANCES+=($(gcloud compute instances list --zones "${ZONE}" --filter="${LABELS}" --format="value(name)" 2>/dev/null))
+fi
+
+echo -e "\nRemoving '${#INSTANCES[@]}' instances"
+
+if [[ ${#INSTANCES[@]} -lt 2 ]]; then
+    echo -e "\nNo instances found...\n"
+    exit 0
 fi
 
 for instance in "${INSTANCES[@]}"
 do
-    echo "Removing $instance..."
+    echo -e "\nRemoving $instance..."
+    echo -e "  -- Removing GKE Hub Assignment"
+    gcloud container hub memberships delete ${instance} --quiet --async 2> /dev/null
+    echo -e "  -- Deleting instance"
     gcloud compute instances delete ${instance} --zone ${ZONE} -q
+    echo -e "  -- Done!\n"
 done
